move scrolling logic inside the overlay component
